moved heleper functions to another component
