+++ game/scoreboard/scoreboardLogic.js
@@ -1,3 +1,5 @@
+import {renderScore} from './scoreboardRender'
+import {generateRandomScore, fillScores} from './scoreboardHelpers'
 const btnx = document.getElementById('scoreboardBtn')
 const scoreboard = document.getElementById('scoreboard')
 
@@ -12,25 +14,6 @@ const addScore = (player, score) => {
   };
   getScore.push(newScoreData);
   return setScore(getScore);
-};
-
-const generateRandomScore = () => Math.floor(Math.random() * 1000);
-const fillScores = num => {
-  const nameArr = [
-    "Kent Sauro",
-    "Donald Mitten",
-    "Sherill Mayson",
-    "Wilbur Overton",
-    "Alison Daughtrey",
-    "Coleen Carrington",
-    "Mara Royce",
-    "Vickey Sarris",
-    "Patsy Narducci",
-    "Josue Weitzel"
-  ];
-  for (let i = 0; i < num; i++) {
-    addScore(nameArr[i], generateRandomScore());
-  }
 };
 
 const sortScores = arr => arr.sort((a, b) => {
